Backend Update.
Added a User CRUD Controller, a Category model and CRUD Controller. Initialized the RBAC, added the inicialization of the Admin user. Made minor changes to the DB for better inicialization with the RBAC. Added the creation of the extra table "Utilizador" everytime a user is created. Added role based restrictions to the backend controllers. Cleaned up the backend index, side bars, and the login page.